Cleanup WebCodecs WPT
A lot of WebCodecs tests are using async_test, which lead to time outs
when there is an error, and potentially missed test failures.

Additionally, a lot of code verifying configs, and which calls are
allowed in which codec state are common to many test files. This CL
extracts some of common elements into its own utils file

Bug: 1094096, 1094182
Change-Id: I7dd6c9ab76959ed6adb86fd2e6bf94d21610838a